requirements.txt
tensorflow
keras
numpy
pandas
matplotlib
seaborn
scikit-learn
opencv-python
streamlit
pillow
